+++ tests/perf/deq/perf_deq_push_back.c
@@ -14,11 +14,11 @@ int main(void)
     {
         deq_int c = deq_int_init();
         int elems = TEST_PERF_CHUNKS * run;
-        int t0 = TEST_TIME();
+        long t0 = TEST_TIME();
         for(int elem = 0; elem < elems; elem++)
             deq_int_push_back(&c, rand());
-        int t1 = TEST_TIME();
-        printf("%10d %10d\n", elems, t1 - t0);
+        long t1 = TEST_TIME();
+        printf("%10d %10ld\n", elems, t1 - t0);
         deq_int_free(&c);
     }
 }
