+++ tests/perf/lst/perf_lst_iterate.c
@@ -9,13 +9,13 @@
 int main(int argc, char** argv)
 {
     int silent = 0;
-    int t0 = TEST_TIME();
+    long t0 = TEST_TIME();
     puts(__FILE__);
     srand(0xbeef);
     if (argc >= 2 && argv[1][0] == '-' && argv[1][1] == 's')
         silent = 1;
 #ifdef PERF // calc. startup delay for perf stat -D
-    int t1 = TEST_TIME();
+    long t1 = TEST_TIME();
     printf("-D%d usec\n", t1 - t0);
 #endif
     for(int run = 0; run < TEST_PERF_RUNS; run++)
@@ -31,7 +31,7 @@ int main(int argc, char** argv)
                 sum += *it.ref;
         }
         if (!silent)
-            printf("%10d %10d\n", elems, TEST_TIME() - t0);
+            printf("%10d %10ld\n", elems, TEST_TIME() - t0);
         list_int_free(&c);
     }
 }
