+++ tests/perf/set/perf_set_iterate.cc
@@ -13,11 +13,11 @@ int main()
         int elems = TEST_PERF_CHUNKS * run;
         for(int elem = 0; elem < elems; elem++)
             c.insert(rand() % elems);
-        int t0 = TEST_TIME();
+        long t0 = TEST_TIME();
         volatile int sum = 0;
         for(auto& x : c)
             sum += x;
-        int t1 = TEST_TIME();
-        printf("%10d %10d\n", elems, t1 - t0);
+        long t1 = TEST_TIME();
+        printf("%10d %10ld\n", elems, t1 - t0);
     }
 }
