+++ tests/perf/uset/perf_uset_insert.c
@@ -17,11 +17,11 @@ int main(void)
     {
         uset_int c = uset_int_init(int_hash, int_equal);
         int elems = TEST_PERF_CHUNKS * run;
-        int t0 = TEST_TIME();
+        long t0 = TEST_TIME();
         for(int elem = 0; elem < elems; elem++)
             uset_int_insert(&c, rand() % elems);
-        int t1 = TEST_TIME();
-        printf("%10d %10d\n", elems, t1 - t0);
+        long t1 = TEST_TIME();
+        printf("%10d %10ld\n", elems, t1 - t0);
         uset_int_free(&c);
     }
 }
